pyproject.toml
[project]
name = "tcc-api"
version = "0.1.0"
description = ""
authors = [
    {name = "Sr1515",email = "joao151576@gmail.com"}
]
requires-python = ">=3.10,<4"
dependencies = [
    "djangorestframework (>=3.15.2,<4.0.0)",
    "django (>=5.1.7,<6.0.0)",
    "django-filter (>=25.1,<26.0)",
    "markdown (>=3.7,<4.0)",
    "pillow (>=11.1.0,<12.0.0)",
    "django-environ (>=0.12.0,<0.13.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "pip (>=25.0.1,<26.0.0)",
    "djangorestframework-simplejwt (>=5.5.0,<6.0.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
